set priviliges for managers and admin

+++ app/Http/Controllers/InventoryController.php
@@ -18,8 +18,13 @@ class InventoryController extends Controller
 
     public function index()
     {
+      // inventories for this manager not all inventories
+      // if admin get all inventories
 
-        $inventories = Inventory::all();
+        if(\Gate::allows('view_all',Inventory::class))
+            $inventories = Inventory::all();
+        elseif(\Gate::allows('view_his_own',Inventory::class))
+            $inventories = collect([Auth::user()->inventory]);
 
         return view('inventory.index',compact('inventories'));
     }
@@ -27,12 +32,15 @@ class InventoryController extends Controller
 
     public function create()
     {
+        $this->authorize('create',Inventory::class);
         return view('inventory.create');
     }
 
 
     public function store()
     {
+      $this->authorize('create',Inventory::class);
+
         Inventory::create(request()->validate([
             'name'=>'required'
           ]));
@@ -49,13 +57,15 @@ class InventoryController extends Controller
 
     public function edit(Inventory $inventory)
     {
-
+      $this->authorize('update',Inventory::class);
         return view('inventory.edit',['inventory'=>$inventory]);
     }
 
 
     public function update(Inventory $inventory)
     {
+      $this->authorize('update',Inventory::class);
+
         $inventory->update(
            request()->validate(['name' => 'required'])
          );
@@ -66,6 +76,7 @@ class InventoryController extends Controller
 
     public function destroy(Inventory $inventory)
     {
+        $this->authorize('delete',Inventory::class);
 
         $inventory->delete();
